Catch errors in evolve.sh

+++ evolve.sh
@@ -39,22 +39,41 @@ echo
 build/sr2 mutate $OPTS --usePopulation $STABLE_FILE \
                        --outputFile $NOOB_FILE \
                        --mutationCount $NOOB_POP
+if [ $? != 0 ]; then
+    exit $?
+fi;
 
 build/sr2 measure $OPTS --usePopulation $NOOB_FILE \
                         --controlPopulation $SCREEN1_FILE \
                         --loop $SCREEN1_ITERATIONS
+if [ $? != 0 ]; then
+    exit $?
+fi;
 
 build/sr2 kill $OPTS --usePopulation $NOOB_FILE \
                      --defectiveLevel $SCREEN1_DEFECTIVE \
                      --resetAfter
+if [ $? != 0 ]; then
+    exit $?
+fi;
 
 build/sr2 merge $OPTS --usePopulation $STABLE_FILE \
                       --inputPopulation $NOOB_FILE
+if [ $? != 0 ]; then
+    exit $?
+fi;
 
 build/sr2 measure $OPTS --usePopulation $STABLE_FILE \
                         --controlPopulation $SCREENS_FILE \
                         --loop $SCREENS_ITERATIONS
+if [ $? != 0 ]; then
+    exit $?
+fi;
 
 build/sr2 kill $OPTS --usePopulation $STABLE_FILE \
                      --maxPop $STABLE_POP
+if [ $? != 0 ]; then
+    exit $?
+fi;
+
